refactoring views
